# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OCBP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/ethereum/get/specialties": {
            "get": {
                "tags": [
                    "Ethereum"
                ],
                "summary": "Get Specialties Info",
                "description": "Returns Specialties Info",
                "operationId": "Specialties",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/ethereum/get/doctors-with-spe-regm": {
            "get": {
                "tags": [
                    "Ethereum"
                ],
                "summary": "Get Doctors Info",
                "description": "Returns Doctors Info",
                "operationId": "Doctors",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/ethereum/get/diagnostics": {
            "get": {
                "tags": [
                    "Ethereum"
                ],
                "summary": "Get Diagnostics Info",
                "description": "Returns Diagnostics Info",
                "operationId": "Diagnostics",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/ethereum/get/patient/{patientdoc}/type/{patientdoctype}/information": {
            "get": {
                "tags": [
                    "Ethereum"
                ],
                "summary": "Get Patient Informations",
                "description": "Returns Patient Information",
                "operationId": "initialPatientInfoEthereum",
                "parameters": [
                    {
                        "name": "patientdoc",
                        "in": "path",
                        "description": "Documento del Paciente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "patientdoctype",
                        "in": "path",
                        "description": "Tipo de Documento del Paciente - CC, TI, RC, PE",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
...
        "/api/v1/exitus/get/occupation-with-real-stay": {
            "get": {
                "tags": [
                    "Exitus"
                ],
                "summary": "Get occupation Info With Real Stay",
                "description": "Returns occupation Name",
                "operationId": "occupation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
...
        "/api/v1/macna/patient/{patientdoc}/type/{patientdoctype}/information": {
            "get": {
                "tags": [
                    "Macna"
                ],
                "summary": "Get Patient Informations",
                "description": "Returns Patient Information",
                "operationId": "initialPatientInfo",
                "parameters": [
                    {
                        "name": "patientdoc",
                        "in": "path",
                        "description": "Documento del Paciente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "patientdoctype",
                        "in": "path",
                        "description": "Tipo de Documento del Paciente - CC, TI, RC, PE",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-Authorization": {
                "type": "apiKey",
                "description": "EspartaApiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "X-Authorization": []
        }
    ]
}